Moving API call into a separate file
