+++ src/components/Home.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import type { BlogModel } from '../type/types';
 import { useState, useEffect } from 'react';
+import { fetchBlogList } from '../services/blogService';
 
 export const Home = (): React.JSX.Element => {
   const [blogs, setBlogs] = useState<BlogModel[]>([]);
@@ -36,31 +37,3 @@ export const Home = (): React.JSX.Element => {
 };
 
 export default Home;
-
-import axios from 'axios';
-
-const API_BASE = 'http://localhost:3000';
-
-import type { BlogListResponse } from '../type/types';
-
-const fetchBlogList = async (): Promise<BlogListResponse> => {
-  try {
-    const response = await axios.get<BlogListResponse>(`${API_BASE}/blogs`);
-    return response.data;
-  } catch (error) {
-    return {
-      success: false,
-      message: extractErrorMessage(error),
-      error: extractErrorMessage(error),
-    };
-  }
-};
-
-function extractErrorMessage(error: unknown): string {
-  if (axios.isAxiosError(error)) {
-    const apiError = error.response?.data as { error?: string } | undefined;
-    if (apiError?.error) return apiError.error;
-    return error.message;
-  }
-  return 'Unknown error occurred';
-}
